Time: 31 ms (66.42%), Space: 106.7 MB (81.61%) - LeetHub

+++ 2126-destroying-asteroids/2126-destroying-asteroids.cpp
@@ -6,7 +6,7 @@ public:
         long long currMass=mass;
         for(int i=0;i<n;i++){
             if(ast[i]>currMass)return false;
-            currMass+=(long long)ast[i];
+            currMass+=ast[i];
         }
         return true;
     }
